requirements.txt
langchain
langchain_core
langchain_community
pymupdf
sentence-transformers
python-docx
chromadb
streamlit
fastapi       
python-dotenv
pandas
uvicorn
langchain_google_genai
langchain-chroma
langchain-huggingface
